refactor(nx-cmake): graph processing now utilizes filesToProcess
