+++ packages/nx-cmake/src/createDependencies/createDependencies.ts
@@ -1,31 +1,17 @@
 import type {
     CreateDependencies,
     CreateDependenciesContext,
-    ProjectFileMap,
     ProjectGraphDependencyWithFile,
 } from '@nx/devkit';
 import { filterProjects } from '../utils/graphUtils/filterProjects/filterProjects';
 import { getDependencies } from '../utils/graphUtils/getDependencies/getDependencies';
 import { reduceDependenciesTransitively } from '../utils/graphUtils/reduceDependenciesTransitively/reduceDependenciesTransitively';
-import { FilteredProject } from '../models/types';
-
-export const filterFilesToProcess = (
-    filesToProcess: ProjectFileMap,
-    filteredProjects: FilteredProject[]
-): ProjectFileMap => {
-    const filteredFilesToProcess: ProjectFileMap = {};
-    for (const { name } of filteredProjects) {
-        if (filesToProcess[name]) {
-            filteredFilesToProcess[name] = filesToProcess[name];
-        }
-    }
-    return filteredFilesToProcess;
-};
+import { filterFilesToProcess } from '../utils/graphUtils/filterFilesToProcess/filterFilesToProcess';
 
 export const createDependencies: CreateDependencies = (
     context: CreateDependenciesContext
 ): ProjectGraphDependencyWithFile[] => {
-    const { graph, nxJsonConfiguration, fileMap, filesToProcess } = context;
+    const { graph, nxJsonConfiguration, filesToProcess } = context;
     if (Object.keys(filesToProcess).length === 0) {
         return [];
     }
@@ -42,7 +28,6 @@ export const createDependencies: CreateDependencies = (
     const deps = getDependencies(
         workspaceLayout,
         filteredProjects,
-        fileMap,
         filteredFilesToProcess
     );
     if (deps.length === 0) {
