+++ packages/nx-cmake/src/utils/graphUtils/filterFilesToProcess/filterFilesToProcess.spec.ts
@@ -1,0 +1,6 @@
+import type { ProjectFileMap } from '@nx/devkit';
+import type { FilteredProject } from '../../../models/types';
+
+describe('filterFilesToProcess', () => {
+    it.todo('should filter files to process');
+});
