+++ packages/nx-cmake/src/utils/graphUtils/getDependencies/getDependencies.spec.ts
@@ -9,7 +9,6 @@ import { getDependencies } from './getDependencies';
 import * as filterDependenciesOfProjectModule from '../filterDependenciesOfProject/filterDependenciesOfProject';
 
 describe('getDependencies', () => {
-    let fileMap: ProjectFileMap;
     let filesToProcess: ProjectFileMap;
     let projects: FilteredProject[];
     let workspaceLayout: WorkspaceLayout;
@@ -71,71 +70,6 @@ describe('getDependencies', () => {
                 dependencyType: DependencyType.static,
             },
         ];
-        fileMap = {
-            parser: [
-                {
-                    file: 'bin/parser/CMakeLists.txt',
-                    hash: '17299240877484337829',
-                },
-                { file: 'bin/parser/README.md', hash: '7643242783975375151' },
-                {
-                    file: 'bin/parser/include/parser.h',
-                    hash: '11018423326186667389',
-                },
-                {
-                    file: 'bin/parser/project.json',
-                    hash: '2184246761215036098',
-                },
-                {
-                    file: 'bin/parser/src/parser.c',
-                    hash: '3553739439408751796',
-                },
-            ],
-            libparser: [
-                {
-                    file: 'packages/parser/CMakeLists.txt',
-                    hash: '4187171820888433546',
-                },
-                {
-                    file: 'packages/parser/README.md',
-                    hash: '519973881722473310',
-                },
-                {
-                    file: 'packages/parser/include/libparser.h',
-                    hash: '9050508492991185026',
-                },
-                {
-                    file: 'packages/parser/project.json',
-                    hash: '3459988573673696231',
-                },
-                {
-                    file: 'packages/parser/src/libparser.c',
-                    hash: '5782631378711784791',
-                },
-            ],
-            testparser: [
-                {
-                    file: 'packages/parser/test/CMakeLists.txt',
-                    hash: '9995994963779964934',
-                },
-                {
-                    file: 'packages/parser/test/README.md',
-                    hash: '13416536286541011897',
-                },
-                {
-                    file: 'packages/parser/test/include/testparser.h',
-                    hash: '9147925740844481238',
-                },
-                {
-                    file: 'packages/parser/test/project.json',
-                    hash: '3367775946370084247',
-                },
-                {
-                    file: 'packages/parser/test/src/testparser.c',
-                    hash: '2754802812570811200',
-                },
-            ],
-        };
         filesToProcess = {};
         filterDependenciesOfProjectMock = jest.spyOn(
             filterDependenciesOfProjectModule,
@@ -180,7 +114,6 @@ describe('getDependencies', () => {
         const result = getDependencies(
             workspaceLayout,
             projects,
-            fileMap,
             filesToProcess
         );
         expect(filterDependenciesOfProjectMock).toBeCalledTimes(0);
